Fix bug 3
- Added check to passes to make Q only available for the player's fourth move if it isn't played yet

+++ main/index.php
@@ -129,8 +129,13 @@
         <form method="post" action="play.php">
             <select name="piece">
                 <?php
-                    foreach ($hand[$player] as $tile => $ct) {
-                        echo "<option value=\"$tile\">$tile</option>";
+                    if(array_sum($hand[$player]) <= 8 && array_key_exists("Q", $hand[$player])){
+                        echo "<option value='Q'>Q</option>";
+                    }
+                    else {
+                        foreach ($hand[$player] as $tile => $ct) {
+                            echo "<option value=\"$tile\">$tile</option>";
+                        }
                     }
                 ?>
             </select>
